requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
supabase>=2.0.0
pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.50.0
